+++ frontend/tailwind.config.js
@@ -6,6 +6,20 @@ export default {
   ],
   theme: {
     extend: {},
+    '2xl': {'max': '1535px'},
+      // => @media (max-width: 1535px) { ... }
+
+      'xl': {'max': '1279px'},
+      // => @media (max-width: 1279px) { ... }
+
+      'lg': {'max': '1023px'},
+      // => @media (max-width: 1023px) { ... }
+
+      'md': {'max': '767px'},
+      // => @media (max-width: 767px) { ... }
+
+      'sm': {'max': '576px'},
+      // => @media (max-width: 576px) { ... }
   },
   plugins: [],
 }
